requirements.txt
allure-pytest==2.8.6
pytest==4.5.0